Move IErrorLog and IPropertyBag interfaces to interop
