+++ src/Common/src/Interop/Ole32/Interop.IErrorLog.cs
@@ -1,0 +1,23 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Runtime.InteropServices;
+
+internal partial class Interop
+{
+    internal static partial class Ole32
+    {
+        [ComImport]
+        [Guid("3127CA40-446E-11CE-8135-00AA004BB851")]
+        [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+        public unsafe interface IErrorLog
+        {
+            [PreserveSig]
+            HRESULT AddError(
+                [MarshalAs(UnmanagedType.LPWStr)] string pszPropName,
+                EXCEPINFO* pExcepInfo);
+        }
+    }
+}
